+++ src/http-router/contracts/http-req.contract.ts
@@ -53,30 +53,7 @@ export type HttpReqFiles<TReqFiles extends FileInputs> = {
     [K in keyof TReqFiles]: IHttpFileCollection;
 };
 
-/**
- * Represents an incoming HTTP request with typed access to all data sources.
- *
- * Provides methods to access parsed and raw values for the body, form fields,
- * path parameters, query parameters, and headers.
- * Cookies are handled separately via {@link ICookieStore}.
- *
- * Each generic type parameter corresponds to a single source of request data,
- * allowing granular type inference without needing a wrapper type.
- *
- * @typeParam TReqMethod - The HTTP method type.
- * @typeParam TReqJson - The type of the parsed JSON body.
- * @typeParam TReqFields - The type of the parsed form fields.
- * @typeParam TReqParams - The type of the parsed path parameters.
- * @typeParam TReqSearchParams - The type of the parsed query parameters.
- * @typeParam TReqHeaders - The type of the parsed headers.
- * @typeParam TReqFiles - The expected file upload definitions.
- * @typeParam TCookieData - A record mapping cookie names to their value types.
- *
- * IMPORT_PATH: `"@daiso-tech/core/http-router/contracts"`
- * @group Contracts
- */
-export type IHttpReq<
-    TReqMethod = HttpMethod,
+export type IValidatedHttpReq<
     TReqJson = unknown,
     TReqFields extends ReqInputs = ReqInputs,
     TReqParams extends ReqInputs = ReqInputs,
@@ -84,31 +61,7 @@ export type IHttpReq<
     TReqHeaders extends ReqInputs = ReqInputs,
     TReqFiles extends FileInputs = FileInputs,
     TCookieData extends StringInputs = StringInputs,
-> = AsyncIterable<unknown> & {
-    /**
-     * An {@link https://developer.mozilla.org/en-US/docs/Web/API/AbortSignal | AbortSignal}
-     * that is aborted when the underlying request is cancelled or times out.
-     *
-     * Handlers can listen to this signal to stop long-running work early when
-     * the client disconnects or the request reaches its deadline.
-     */
-    readonly signal: AbortSignal;
-
-    /**
-     * Returns the HTTP method of the request (e.g. GET, POST).
-     */
-    readonly method: TReqMethod;
-
-    /**
-     * Returns the full request URL as a string.
-     */
-    readonly url: string;
-
-    /**
-     * Returns the raw, parsed cookies as a string-to-string map.
-     */
-    rawCookies(): StringInputs;
-
+> = {
     /**
      * Parses and returns all cookies, validated and transformed by the cookie schema.
      */
@@ -133,11 +86,6 @@ export type IHttpReq<
      * Parses and returns the JSON body.
      */
     json(): Promise<TReqJson>;
-
-    /**
-     * Returns the raw unparsed JSON body.
-     */
-    rawJson(): Promise<unknown>;
 
     /**
      * Parses and returns all form fields from `Request.formData`.
@@ -170,12 +118,6 @@ export type IHttpReq<
     ): Promise<IHttpFileCollection>;
 
     /**
-     * Returns the raw unparsed `FormData` as a plain object, or an empty object if not present.
-     * Each field is either a string value or an {@link IHttpFile} instance for file uploads.
-     */
-    rawFormData(): Promise<RawFormData>;
-
-    /**
      * Parses and returns all path parameters.
      */
     params(): TReqParams;
@@ -189,11 +131,6 @@ export type IHttpReq<
     params<TField extends keyof TReqParams, TValue extends TReqParams[TField]>(
         field: TField,
     ): UndefinedToNull<TValue>;
-
-    /**
-     * Returns the raw unparsed path parameters, or empty object if not present.
-     */
-    rawParams(): StringInputs;
 
     /**
      * Parses and returns all query string parameters.
@@ -214,11 +151,6 @@ export type IHttpReq<
     ): UndefinedToNull<TValue>;
 
     /**
-     * Returns the raw unparsed query parameters, or empty object if not present.
-     */
-    rawSearchParams(): MultiStringInputs;
-
-    /**
      * Parses and returns all request headers.
      */
     headers(): TReqHeaders;
@@ -235,6 +167,146 @@ export type IHttpReq<
     >(
         field: TField,
     ): UndefinedToNull<TValue>;
+};
+
+/**
+ * Defines optional validation schemas for each source of request data.
+ * Each field uses Standard Schema V1 to validate and transform the input.
+ * Cookies are not validated here but handled separately via {@link ICookieStore}.
+ *
+ * Each field maps directly to its own type parameter so TypeScript can infer
+ * the output type from the schema (e.g. from a Zod schema) without needing
+ * to reverse through indexed access types.
+ *
+ * @typeParam TReqJson - The type of the parsed JSON body.
+ * @typeParam TReqFields - The type of the parsed form fields.
+ * @typeParam TReqParams - The type of the parsed path parameters.
+ * @typeParam TReqSearchParams - The type of the parsed query parameters.
+ * @typeParam TReqHeaders - The type of the parsed headers.
+ * @typeParam TReqFiles - The expected file definitions.
+ *
+ * IMPORT_PATH: `"@daiso-tech/core/http-router/contracts"`
+ * @group Contracts
+ */
+export type HttpReqSchemas<
+    TReqJson = unknown,
+    TReqFields extends ReqInputs = ReqInputs,
+    TReqParams extends ReqInputs = ReqInputs,
+    TReqSearchParams extends ReqInputs = ReqInputs,
+    TReqHeaders extends ReqInputs = ReqInputs,
+    TReqFiles extends FileInputs = FileInputs,
+    TCookieData extends StringInputs = StringInputs,
+> = {
+    /**
+     * Optional schema for validating the JSON body.
+     */
+    json?: StandardSchemaV1<TReqJson>;
+
+    /**
+     * Optional schema for validating form fields or url encoded fields.
+     * Receives {@link MultiStringInputs} (string or string[]) since form fields
+     * may appear multiple times.
+     */
+    fields?: StandardSchemaV1<MultiStringInputs, TReqFields>;
+
+    /**
+     * Optional expected file definitions for validation.
+     */
+    files?: TReqFiles;
+
+    /**
+     * Optional schema for validating path parameters.
+     */
+    params?: StandardSchemaV1<StringInputs, TReqParams>;
+
+    /**
+     * Optional schema for validating query string parameters.
+     */
+    searchParams?: StandardSchemaV1<MultiStringInputs, TReqSearchParams>;
+
+    /**
+     * Optional schema for validating request headers.
+     */
+    headers?: StandardSchemaV1<StringInputs, TReqHeaders>;
+
+    cookies?: StandardSchemaV1<StringInputs, TCookieData>;
+};
+
+export type IHttpReqValidation = {
+    withSchema<
+        TReqJson = unknown,
+        TReqFields extends ReqInputs = ReqInputs,
+        TReqParams extends ReqInputs = ReqInputs,
+        TReqSearchParams extends ReqInputs = ReqInputs,
+        TReqHeaders extends ReqInputs = ReqInputs,
+        TReqFiles extends FileInputs = FileInputs,
+        TCookieData extends StringInputs = StringInputs,
+    >(
+        schemas: HttpReqSchemas<
+            TReqJson,
+            TReqFields,
+            TReqParams,
+            TReqSearchParams,
+            TReqHeaders,
+            TReqFiles,
+            TCookieData
+        >,
+    ): IValidatedHttpReq<
+        TReqJson,
+        TReqFields,
+        TReqParams,
+        TReqSearchParams,
+        TReqHeaders,
+        TReqFiles,
+        TCookieData
+    >;
+};
+
+export type IHttpReqBase = AsyncIterable<unknown> & {
+    /**
+     * An {@link https://developer.mozilla.org/en-US/docs/Web/API/AbortSignal | AbortSignal}
+     * that is aborted when the underlying request is cancelled or times out.
+     *
+     * Handlers can listen to this signal to stop long-running work early when
+     * the client disconnects or the request reaches its deadline.
+     */
+    readonly signal: AbortSignal;
+
+    /**
+     * Returns the HTTP method of the request (e.g. GET, POST).
+     */
+    readonly method: HttpMethod;
+
+    /**
+     * Returns the full request URL as a string.
+     */
+    readonly url: string;
+
+    /**
+     * Returns the raw, parsed cookies as a string-to-string map.
+     */
+    rawCookies(): StringInputs;
+
+    /**
+     * Returns the raw unparsed JSON body.
+     */
+    rawJson(): Promise<unknown>;
+
+    /**
+     * Returns the raw unparsed `FormData` as a plain object, or an empty object if not present.
+     * Each field is either a string value or an {@link IHttpFile} instance for file uploads.
+     */
+    rawFormData(): Promise<RawFormData>;
+
+    /**
+     * Returns the raw unparsed path parameters, or empty object if not present.
+     */
+    rawParams(): StringInputs;
+
+    /**
+     * Returns the raw unparsed query parameters, or empty object if not present.
+     */
+    rawSearchParams(): MultiStringInputs;
 
     /**
      * Returns the raw unparsed headers, or empty object if not present.
@@ -280,61 +352,24 @@ export type IHttpReq<
 };
 
 /**
- * Defines optional validation schemas for each source of request data.
- * Each field uses Standard Schema V1 to validate and transform the input.
- * Cookies are not validated here but handled separately via {@link ICookieStore}.
+ * Represents an incoming HTTP request with typed access to all data sources.
  *
- * Each field maps directly to its own type parameter so TypeScript can infer
- * the output type from the schema (e.g. from a Zod schema) without needing
- * to reverse through indexed access types.
+ * Provides methods to access parsed and raw values for the body, form fields,
+ * path parameters, query parameters, and headers.
+ * Cookies are handled separately via {@link ICookieStore}.
+ *
+ * Each generic type parameter corresponds to a single source of request data,
+ * allowing granular type inference without needing a wrapper type.
  *
  * @typeParam TReqJson - The type of the parsed JSON body.
  * @typeParam TReqFields - The type of the parsed form fields.
  * @typeParam TReqParams - The type of the parsed path parameters.
  * @typeParam TReqSearchParams - The type of the parsed query parameters.
  * @typeParam TReqHeaders - The type of the parsed headers.
- * @typeParam TReqFiles - The expected file definitions.
+ * @typeParam TReqFiles - The expected file upload definitions.
+ * @typeParam TCookieData - A record mapping cookie names to their value types.
  *
  * IMPORT_PATH: `"@daiso-tech/core/http-router/contracts"`
  * @group Contracts
  */
-export type HttpReqSchema<
-    TReqJson = unknown,
-    TReqFields extends ReqInputs = ReqInputs,
-    TReqParams extends ReqInputs = ReqInputs,
-    TReqSearchParams extends ReqInputs = ReqInputs,
-    TReqHeaders extends ReqInputs = ReqInputs,
-    TReqFiles extends FileInputs = FileInputs,
-> = {
-    /**
-     * Optional schema for validating the JSON body.
-     */
-    json?: StandardSchemaV1<TReqJson>;
-
-    /**
-     * Optional schema for validating form fields or url encoded fields.
-     * Receives {@link MultiStringInputs} (string or string[]) since form fields
-     * may appear multiple times.
-     */
-    fields?: StandardSchemaV1<MultiStringInputs, TReqFields>;
-
-    /**
-     * Optional expected file definitions for validation.
-     */
-    files?: TReqFiles;
-
-    /**
-     * Optional schema for validating path parameters.
-     */
-    params?: StandardSchemaV1<StringInputs, TReqParams>;
-
-    /**
-     * Optional schema for validating query string parameters.
-     */
-    searchParams?: StandardSchemaV1<MultiStringInputs, TReqSearchParams>;
-
-    /**
-     * Optional schema for validating request headers.
-     */
-    headers?: StandardSchemaV1<StringInputs, TReqHeaders>;
-};
+export type IHttpReq = IHttpReqValidation & IHttpReqBase;
